Moved stt handler to stt folder created whisper class from base stt_handler
